refactor to split out class
